requirements.txt
python-telegram-bot>=20
fastapi
uvicorn[standard]
